workload identity: add claim for namespace job ID
When defining Vault entities the `user_claim` must be unique. When writing Vault
binding rules for use with Nomad workload identities the binding rule won't be
able to create a 1:1 mapping because the selector language allows accessing only
a single field. The `nomad_job_id` claim isn't sufficient to uniquely identify a
job because of namespaces. It's possible to create a JWT auth role with
`bound_claims` to avoid this becoming a security problem, but this doesn't allow
for correct accounting of user claims.

Add a new claim `nomad_workload_id` that uniquely identifies a Nomad job by
using the namespaced job ID (with a separator that cannot appear inside a
namespace name). This will allow any external consumer of WI to use a single
claim field for binding rules, so long as that consumer is ok with sharing the
binding rule across groups within a job or tasks within a group (at which point
they'll need to go look at the task/service fields).

Fixes: https://github.com/hashicorp/nomad/issues/23510
Ref: https://hashicorp.atlassian.net/browse/NET-10372
Ref: https://hashicorp.atlassian.net/browse/NET-10387