E2E: add Vault compatibility test demonstrating entity alias use
Although we encourage users to use Vault roles, sometimes they're going to want
to assign policies based on entity and precreate entities and aliases based on
claims like `nomad_workload_id`. This allows them to use single default role (or at least
small number of them) that has a templated policy, but have an escape hatch from
that.

This commit includes a test of that workflow.